moved constants to constants folder
