+++ client/src/gameData/otherConstants.js
@@ -1,5 +1,3 @@
-//TODO: could move into master constant file
-
 export const SOCKET_SERVER_REDIRECT = "serverRedirect";
 export const SOCKET_SERVER_SENDING_ACTION = "serverSendingAction";
 export const SOCKET_CLIENT_SENDING_ACTION = "clientSendingAction";
