+++ client/src/redux/actions/capabilities/atcScramble.js
@@ -1,10 +1,9 @@
 import setUserfeedbackAction from "../setUserfeedbackAction";
-// import { CLIENT_SENDING_ACTION } from "../../socketEmits";
 
 const atcScramble = invItem => {
-	return (dispatch, getState, emit) => {
-		dispatch(setUserfeedbackAction("atcScramble"));
-	};
+    return (dispatch, getState, emit) => {
+        dispatch(setUserfeedbackAction("atcScramble"));
+    };
 };
 
 export default atcScramble;
