+++ client/src/redux/actions/capabilities/cyberDominance.js
@@ -1,10 +1,9 @@
 import setUserfeedbackAction from "../setUserfeedbackAction";
-// import { CLIENT_SENDING_ACTION } from "../../socketEmits";
 
 const cyberDominance = invItem => {
-	return (dispatch, getState, emit) => {
-		dispatch(setUserfeedbackAction("cyberDominance"));
-	};
+    return (dispatch, getState, emit) => {
+        dispatch(setUserfeedbackAction("cyberDominance"));
+    };
 };
 
 export default cyberDominance;
