+++ client/src/redux/actions/capabilities/droneSwarms.js
@@ -1,10 +1,9 @@
 import setUserfeedbackAction from "../setUserfeedbackAction";
-// import { CLIENT_SENDING_ACTION } from "../../socketEmits";
 
 const droneSwarms = invItem => {
-	return (dispatch, getState, emit) => {
-		dispatch(setUserfeedbackAction("droneSwarms"));
-	};
+    return (dispatch, getState, emit) => {
+        dispatch(setUserfeedbackAction("droneSwarms"));
+    };
 };
 
 export default droneSwarms;
