+++ client/src/redux/actions/capabilities/missileLaunchDisruption.js
@@ -1,10 +1,9 @@
 import setUserfeedbackAction from "../setUserfeedbackAction";
-// import { CLIENT_SENDING_ACTION } from "../../socketEmits";
 
 const missileLaunchDisruption = invItem => {
-	return (dispatch, getState, emit) => {
-		dispatch(setUserfeedbackAction("missileLaunchDisruption"));
-	};
+    return (dispatch, getState, emit) => {
+        dispatch(setUserfeedbackAction("missileLaunchDisruption"));
+    };
 };
 
 export default missileLaunchDisruption;
