+++ client/src/redux/actions/capabilities/nuclearStrike.js
@@ -1,10 +1,9 @@
 import setUserfeedbackAction from "../setUserfeedbackAction";
-// import { CLIENT_SENDING_ACTION } from "../../socketEmits";
 
 const nuclearStrike = invItem => {
-	return (dispatch, getState, emit) => {
-		dispatch(setUserfeedbackAction("nuclearStrike"));
-	};
+    return (dispatch, getState, emit) => {
+        dispatch(setUserfeedbackAction("nuclearStrike"));
+    };
 };
 
 export default nuclearStrike;
