+++ client/src/redux/actions/capabilities/seaMines.js
@@ -1,10 +1,9 @@
 import setUserfeedbackAction from "../setUserfeedbackAction";
-// import { CLIENT_SENDING_ACTION } from "../../socketEmits";
 
 const seaMines = invItem => {
-	return (dispatch, getState, emit) => {
-		dispatch(setUserfeedbackAction("seaMines"));
-	};
+    return (dispatch, getState, emit) => {
+        dispatch(setUserfeedbackAction("seaMines"));
+    };
 };
 
 export default seaMines;
