+++ client/src/redux/actions/planning/confirmPlan.js
@@ -1,25 +1,25 @@
 import setUserfeedbackAction from "../setUserfeedbackAction";
-import { CLIENT_SENDING_ACTION } from "../../socketEmits";
+import { SOCKET_CLIENT_SENDING_ACTION } from "../../../gameData/otherConstants";
 import { SERVER_CONFIRM_PLAN } from "../actionTypes";
 
 const confirmPlan = () => {
-	return (dispatch, getState, emit) => {
-		const { gameboardMeta } = getState();
+    return (dispatch, getState, emit) => {
+        const { gameboardMeta } = getState();
 
-		if (gameboardMeta.planning.moves.length === 0) {
-			dispatch(setUserfeedbackAction("Can't submit an empty plan..."));
-		} else {
-			const clientAction = {
-				type: SERVER_CONFIRM_PLAN,
-				payload: {
-					pieceId: gameboardMeta.selectedPiece.pieceId,
-					plan: gameboardMeta.planning.moves
-				}
-			};
+        if (gameboardMeta.planning.moves.length === 0) {
+            dispatch(setUserfeedbackAction("Can't submit an empty plan..."));
+        } else {
+            const clientAction = {
+                type: SERVER_CONFIRM_PLAN,
+                payload: {
+                    pieceId: gameboardMeta.selectedPiece.pieceId,
+                    plan: gameboardMeta.planning.moves
+                }
+            };
 
-			emit(CLIENT_SENDING_ACTION, clientAction);
-		}
-	};
+            emit(SOCKET_CLIENT_SENDING_ACTION, clientAction);
+        }
+    };
 };
 
 export default confirmPlan;
