+++ client/src/redux/actions/shop/shopConfirmPurchase.js
@@ -1,4 +1,4 @@
-import { CLIENT_SENDING_ACTION } from "../../socketEmits";
+import { SOCKET_CLIENT_SENDING_ACTION } from "../../../gameData/otherConstants";
 import { SERVER_SHOP_CONFIRM_PURCHASE } from "../actionTypes";
 
 //TODO: more checks on if they can purchase before sending to backend (don't waste network)
@@ -8,7 +8,7 @@ const shopConfirmPurchase = () => {
             type: SERVER_SHOP_CONFIRM_PURCHASE,
             payload: {}
         };
-        emit(CLIENT_SENDING_ACTION, clientAction);
+        emit(SOCKET_CLIENT_SENDING_ACTION, clientAction);
     };
 };
