+++ client/src/redux/actions/shop/shopPurchaseRequest.js
@@ -1,16 +1,16 @@
-import { CLIENT_SENDING_ACTION } from "../../socketEmits";
+import { SOCKET_CLIENT_SENDING_ACTION } from "../../../gameData/otherConstants";
 import { SERVER_SHOP_PURCHASE_REQUEST } from "../actionTypes";
 
 const shopPurchaseRequest = shopItemTypeId => {
-	return (dispatch, getState, emit) => {
-		const clientAction = {
-			type: SERVER_SHOP_PURCHASE_REQUEST,
-			payload: {
-				shopItemTypeId
-			}
-		};
-		emit(CLIENT_SENDING_ACTION, clientAction);
-	};
+    return (dispatch, getState, emit) => {
+        const clientAction = {
+            type: SERVER_SHOP_PURCHASE_REQUEST,
+            payload: {
+                shopItemTypeId
+            }
+        };
+        emit(SOCKET_CLIENT_SENDING_ACTION, clientAction);
+    };
 };
 
 export default shopPurchaseRequest;
