+++ client/src/redux/actions/shop/shopRefundRequest.js
@@ -1,16 +1,16 @@
-import { CLIENT_SENDING_ACTION } from "../../socketEmits";
+import { SOCKET_CLIENT_SENDING_ACTION } from "../../../gameData/otherConstants";
 import { SERVER_SHOP_REFUND_REQUEST } from "../actionTypes";
 
 const shopRefundRequest = shopItem => {
-	return (dispatch, getState, emit) => {
-		const clientAction = {
-			type: SERVER_SHOP_REFUND_REQUEST,
-			payload: {
-				shopItem
-			}
-		};
-		emit(CLIENT_SENDING_ACTION, clientAction);
-	};
+    return (dispatch, getState, emit) => {
+        const clientAction = {
+            type: SERVER_SHOP_REFUND_REQUEST,
+            payload: {
+                shopItem
+            }
+        };
+        emit(SOCKET_CLIENT_SENDING_ACTION, clientAction);
+    };
 };
 
 export default shopRefundRequest;
