+++ server/actions/capabilities/insurgencyConfirm.js
@@ -1,6 +1,6 @@
 const { Game, InvItem, Capability } = require("../../classes");
 import { INSURGENCY_SELECTED } from "../../../client/src/redux/actions/actionTypes";
-import { SERVER_REDIRECT, SERVER_SENDING_ACTION } from "../../../client/src/redux/socketEmits";
+import { SOCKET_SERVER_REDIRECT, SOCKET_SERVER_SENDING_ACTION } from "../../../client/src/gameData/otherConstants";
 import { GAME_INACTIVE_TAG, GAME_DOES_NOT_EXIST } from "../../pages/errorTypes";
 import { INSURGENCY_TYPE_ID, COMBAT_PHASE_ID, SLICE_PLANNING_ID, TYPE_MAIN } from "../../../client/src/gameData/gameConstants";
 const sendUserFeedback = require("../sendUserFeedback");
@@ -17,14 +17,14 @@ const insurgencyConfirm = async (socket, payload) => {
 
     const thisGame = await new Game({ gameId }).init();
     if (!thisGame) {
-        socket.emit(SERVER_REDIRECT, GAME_DOES_NOT_EXIST);
+        socket.emit(SOCKET_SERVER_REDIRECT, GAME_DOES_NOT_EXIST);
         return;
     }
 
     const { gameActive, gamePhase, gameSlice, game0Points, game1Points } = thisGame;
 
     if (!gameActive) {
-        socket.emit(SERVER_REDIRECT, GAME_INACTIVE_TAG);
+        socket.emit(SOCKET_SERVER_REDIRECT, GAME_INACTIVE_TAG);
         return;
     }
 
@@ -84,7 +84,7 @@ const insurgencyConfirm = async (socket, payload) => {
             selectedPositionId
         }
     };
-    socket.emit(SERVER_SENDING_ACTION, serverAction);
+    socket.emit(SOCKET_SERVER_SENDING_ACTION, serverAction);
 };
 
 module.exports = insurgencyConfirm;
