+++ server/actions/planning/confirmPlan.js
@@ -1,7 +1,7 @@
 const { Game, Piece, Plan } = require("../../classes");
 const sendUserFeedback = require("../sendUserFeedback");
 import { PLAN_WAS_CONFIRMED } from "../../../client/src/redux/actions/actionTypes";
-import { SERVER_REDIRECT, SERVER_SENDING_ACTION } from "../../../client/src/redux/socketEmits";
+import { SOCKET_SERVER_REDIRECT, SOCKET_SERVER_SENDING_ACTION } from "../../../client/src/gameData/otherConstants";
 import { GAME_INACTIVE_TAG } from "../../pages/errorTypes";
 import { CONTAINER_TYPES, COMBAT_PHASE_ID, SLICE_PLANNING_ID, TYPE_TERRAIN, TYPE_OWNERS } from "../../../client/src/gameData/gameConstants";
 import { distanceMatrix } from "../../../client/src/gameData/distanceMatrix";
@@ -15,7 +15,7 @@ const confirmPlan = async (socket, payload) => {
     const { gameActive, gamePhase, gameSlice } = thisGame;
 
     if (!gameActive) {
-        socket.emit(SERVER_REDIRECT, GAME_INACTIVE_TAG);
+        socket.emit(SOCKET_SERVER_REDIRECT, GAME_INACTIVE_TAG);
         return;
     }
 
@@ -126,7 +126,7 @@ const confirmPlan = async (socket, payload) => {
             plan
         }
     };
-    socket.emit(SERVER_SENDING_ACTION, serverAction);
+    socket.emit(SOCKET_SERVER_SENDING_ACTION, serverAction);
 };
 
 module.exports = confirmPlan;
