+++ server/actions/planning/deletePlan.js
@@ -1,7 +1,7 @@
 const { Game, Piece } = require("../../classes");
 const sendUserFeedback = require("../sendUserFeedback");
 import { DELETE_PLAN } from "../../../client/src/redux/actions/actionTypes";
-import { SERVER_REDIRECT, SERVER_SENDING_ACTION } from "../../../client/src/redux/socketEmits";
+import { SOCKET_SERVER_REDIRECT, SOCKET_SERVER_SENDING_ACTION } from "../../../client/src/gameData/otherConstants";
 import { GAME_INACTIVE_TAG } from "../../pages/errorTypes";
 import { COMBAT_PHASE_ID, SLICE_PLANNING_ID } from "../../../client/src/gameData/gameConstants";
 
@@ -13,7 +13,7 @@ const deletePlan = async (socket, payload) => {
     const { gameActive, gamePhase, gameSlice } = thisGame;
 
     if (!gameActive) {
-        socket.emit(SERVER_REDIRECT, GAME_INACTIVE_TAG);
+        socket.emit(SOCKET_SERVER_REDIRECT, GAME_INACTIVE_TAG);
         return;
     }
 
@@ -45,7 +45,7 @@ const deletePlan = async (socket, payload) => {
             pieceId
         }
     };
-    socket.emit(SERVER_SENDING_ACTION, serverAction); //TODO: should the other sockets for this team get the update? (in the background?)
+    socket.emit(SOCKET_SERVER_SENDING_ACTION, serverAction); //TODO: should the other sockets for this team get the update? (in the background?)
 };
 
 module.exports = deletePlan;
