+++ server/actions/sendUserFeedback.js
@@ -1,14 +1,14 @@
 import { SET_USERFEEDBACK } from "../../client/src/redux/actions/actionTypes";
-import { SERVER_SENDING_ACTION } from "../../client/src/redux/socketEmits";
+import { SOCKET_SERVER_SENDING_ACTION } from "../../client/src/gameData/otherConstants";
 
 const sendUserFeedback = async (socket, userFeedback) => {
-	const serverAction = {
-		type: SET_USERFEEDBACK,
-		payload: {
-			userFeedback
-		}
-	};
-	socket.emit(SERVER_SENDING_ACTION, serverAction);
+    const serverAction = {
+        type: SET_USERFEEDBACK,
+        payload: {
+            userFeedback
+        }
+    };
+    socket.emit(SOCKET_SERVER_SENDING_ACTION, serverAction);
 };
 
 module.exports = sendUserFeedback;
